Lab_1/report_manager/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pymongo==4.5.0
python-dotenv==1.0.0
grpcio==1.59.0
grpcio-tools==1.59.0
requests==2.31.0
redis==5.0.1
urllib3==2.0.7